requirements.txt
asgiref==3.11.0
django==4.2.27
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.5
typing-extensions==4.15.0
tzdata==2025.3
